requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.8.1
colorama==0.4.1
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.7
dask==1.1.4
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-allauth==0.39.1
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-filter==2.1.0
django-heroku==0.3.1
django-jsonfield==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
docopt==0.6.2
docutils==0.14
drf-jwt-devices==1.2.1
entrypoints==0.2.3
gast==0.2.0
grpcio==1.17.0
gunicorn==19.9.0
h5py==2.8.0
idna==2.8
imutils==0.5.2
input-reader==1.2.2
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonfield==2.0.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Kivy==1.10.1
Kivy-examples==1.10.1
Kivy-Garden==0.1.4
kivy.deps.glew==0.1.10
kivy.deps.gstreamer==0.1.13
kivy.deps.sdl2==0.1.18
kiwisolver==1.0.1
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mock==3.0.5
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.2
numpy==1.16.3
oauthlib==3.0.2
object-detection==0.1
opencv-python==4.0.0.21
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.4.1
pipenv==2018.11.26
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
protobuf-converter==0.2.4
psycopg2==2.8.3
psycopg2-binary==2.8.3
py-make==0.1.1
pycocotools==2.0
Pygments==2.3.0
PyJWT==1.7.1
pyparsing==2.3.1
pypiwin32==223
PyQt5==5.12.1
PyQt5-sip==4.19.15
PySide2==5.12.2
python-dateutil==2.7.5
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2018.9
PyWavelets==1.0.2
pywin32==224
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
regex==2018.11.22
requests==2.21.0
requests-oauthlib==1.2.0
scikit-image==0.14.2
scipy==1.2.1
Send2Trash==1.5.0
shiboken2==5.12.2
six==1.11.0
soupsieve==1.7.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
uber-rides==0.6.0
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.1
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
